[#3] refactor: 권한 관련 예외는 security 패키지에서 관리하도록 변경
