+++ src/main/java/com/fakebook/dongheon/member/service/MemberService.java
@@ -3,8 +3,8 @@ package com.fakebook.dongheon.member.service;
 import com.fakebook.dongheon.member.domain.CustomMemberRepository;
 import com.fakebook.dongheon.member.domain.Member;
 import com.fakebook.dongheon.member.exception.AlreadyExistMemberIdException;
-import com.fakebook.dongheon.member.exception.NotAuthorizedException;
 import com.fakebook.dongheon.member.web.dto.MemberRegisterDto;
+import com.fakebook.dongheon.security.exception.NotAuthorizedException;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
