+++ src/main/java/com/fakebook/dongheon/security/exception/NotAuthorizedException.java
@@ -1,4 +1,4 @@
-package com.fakebook.dongheon.member.exception;
+package com.fakebook.dongheon.security.exception;
 
 public class NotAuthorizedException extends RuntimeException {
 	private static final String message = "잘못된 접근입니다. ";
