+++ src/test/java/com/fakebook/dongheon/member/service/MemberServiceTest.java
@@ -5,8 +5,8 @@ import com.fakebook.dongheon.member.domain.Gender;
 import com.fakebook.dongheon.member.domain.Member;
 import com.fakebook.dongheon.member.exception.AlreadyExistMemberIdException;
 import com.fakebook.dongheon.member.exception.MemberNotFoundException;
-import com.fakebook.dongheon.member.exception.NotAuthorizedException;
 import com.fakebook.dongheon.member.web.dto.MemberRegisterDto;
+import com.fakebook.dongheon.security.exception.NotAuthorizedException;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
